Using 'requests' library and improved 'granted access' logging
The 'requests' library makes error handling much easier, as it follows
the Python principle that every remote error inherits from OSError.
Also, it gives access to the full response from the exception handling.

When an error occurs, the full request/response can still be logged.
This allows showing access was granted, but an exception occurred.